+++ src/Ray.BiliBiliTool.Console/BiliBiliToolHostedService.cs
@@ -43,6 +43,8 @@ namespace Ray.BiliBiliTool.Console
 
         public Task StartAsync(CancellationToken cancellationToken)
         {
+            var isNotifySingle = _configuration.GetSection("Notification:IsSingleAccountSingleNotify").Get<bool>();
+
             try
             {
                 _logger.LogInformation("BiliBiliToolPro 开始运行...{newLine}", Environment.NewLine);
@@ -63,6 +65,11 @@ namespace Ray.BiliBiliTool.Console
                     try
                     {
                         DoTasks(tasks);
+                        if (isNotifySingle)
+                        {
+                            LogAppInfo();
+                            _logger.LogInformation("·开始推送·{task}·{user}", $"{_configuration["RunTasks"]}任务", $"账号【{_cookieStrFactory.CurrentNum}】");
+                        }
                     }
                     catch (Exception e)
                     {
@@ -78,9 +85,11 @@ namespace Ray.BiliBiliTool.Console
             }
             finally
             {
-                LogAppInfo();
-
-                _logger.LogInformation("·开始推送·{task}·{user}", _configuration["RunTasks"],"");
+                if (!isNotifySingle)
+                {
+                    LogAppInfo();
+                    _logger.LogInformation("·开始推送·{task}·{user}", $"{_configuration["RunTasks"]}任务", "");
+                }
                 _logger.LogInformation("运行结束");
                 _applicationLifetime.StopApplication();
             }